Rename iterator files
