+++ tests/CMakeLists.txt
@@ -39,10 +39,10 @@ add_text_ext(_test_compatibility_1)
 add_text_ext(_test_compatibility_2)
 add_text_ext(_test_contextlib)
 add_text_ext(_test_hash)
+add_text_ext(_test_iterator)
 add_text_ext(_test_mapping)
 add_text_ext(_test_mutable_mapping)
 add_text_ext(_test_numpy)
 add_text_ext(_test_py_module_)
-add_text_ext(_test_pybind11)
 add_text_ext(_test_sequence)
 add_text_ext(_test_types)
