+++ tests/_test_iterator.py.cpp
@@ -1,6 +1,6 @@
 #include <pybind11/pybind11.h>
 
-#include <amulet/pybind11_extensions/pybind11.hpp>
+#include <amulet/pybind11_extensions/iterator.hpp>
 
 namespace py = pybind11;
 namespace pyext = Amulet::pybind11_extensions;
@@ -13,7 +13,7 @@ public:
     }
 };
 
-PYBIND11_MODULE(_test_pybind11, m)
+PYBIND11_MODULE(_test_iterator, m)
 {
     m.def("get_iterator", []() { return pyext::make_iterator(MyIterator()); });
 }
